+++ lazarkit/src/main/web/src/lib/ui/widgets/Configurables.svelte
@@ -1,17 +1,8 @@
 <script lang="ts">
   import { info, socket } from "$lib"
-  import { Types, type GenericTypeJson } from "$lib/genericType"
+  import { type GenericTypeJson } from "$lib/genericType"
   import { Section } from "$primitives"
-  import BooleanInput from "$ui/primitives/BooleanInput.svelte"
-  import SelectInput from "$ui/primitives/SelectInput.svelte"
-  import StringInput from "$ui/primitives/StringInput.svelte"
-  import {
-    doubleValidator,
-    floatValidator,
-    intValidator,
-    longValidator,
-    stringValidator,
-  } from "$ui/primitives/validators"
+  import Field from "$ui/Field.svelte"
 
   function processFields(fields: GenericTypeJson[]): {
     [key: string]: GenericTypeJson[]
@@ -42,22 +33,6 @@
   }
   function getStartClassName(item: string): string {
     return item.split("." + getClassName(item))[0]
-  }
-
-  function sendFieldUpdate(item: GenericTypeJson) {
-    if (!item.isValid) return
-    socket.sendMessage({
-      kind: "jvmFields",
-      fields: [
-        {
-          className: item.className,
-          fieldName: item.fieldName,
-          type: item.type,
-          valueString: item.newValueString,
-        },
-      ],
-    })
-    item.valueString = item.newValueString
   }
 
   function sendAllUpdates(fields: GenericTypeJson[]) {
@@ -92,97 +67,7 @@
     <div>
       <h3>{key}</h3>
       {#each items as item}
-        <p>{item.fieldName} {item.type} {item.isValid}</p>
-        {#if item.valueString != item.newValueString && item.isValid}
-          <button
-            onclick={() => {
-              if (!item.isValid) return
-              sendFieldUpdate(item)
-            }}
-          >
-            Update
-          </button>
-        {/if}
-        {#if [Types.INT, Types.LONG, Types.DOUBLE, Types.FLOAT, Types.STRING].includes(item.type)}
-          <form
-            class="item"
-            class:disabled={item.type == Types.UNKNOWN}
-            onsubmit={() => {
-              sendFieldUpdate(item)
-            }}
-          >
-            {#if item.type == Types.INT}
-              <StringInput
-                bind:value={item.value}
-                bind:isValid={item.isValid}
-                bind:startValue={item.valueString}
-                bind:currentValue={item.newValueString}
-                validate={intValidator}
-              />
-            {:else if item.type == Types.LONG}
-              <StringInput
-                bind:value={item.value}
-                bind:isValid={item.isValid}
-                bind:startValue={item.valueString}
-                bind:currentValue={item.newValueString}
-                validate={longValidator}
-                extraChar={"L"}
-              />
-            {:else if item.type == Types.DOUBLE}
-              <StringInput
-                bind:value={item.value}
-                bind:isValid={item.isValid}
-                bind:startValue={item.valueString}
-                bind:currentValue={item.newValueString}
-                validate={doubleValidator}
-              />
-            {:else if item.type == Types.FLOAT}
-              <StringInput
-                bind:value={item.value}
-                bind:isValid={item.isValid}
-                bind:startValue={item.valueString}
-                bind:currentValue={item.newValueString}
-                validate={floatValidator}
-                extraChar={"f"}
-              />
-            {:else if item.type == Types.STRING}
-              <StringInput
-                bind:value={item.value}
-                bind:isValid={item.isValid}
-                bind:startValue={item.valueString}
-                bind:currentValue={item.newValueString}
-                validate={stringValidator}
-              />
-            {:else}
-              {JSON.stringify(item)}
-            {/if}
-          </form>
-        {:else if item.type == Types.BOOLEAN}
-          <BooleanInput
-            bind:value={item.value}
-            bind:isValid={item.isValid}
-            bind:startValue={item.valueString}
-            bind:currentValue={item.newValueString}
-          />
-        {:else if item.type == Types.ENUM}
-          <SelectInput
-            bind:value={item.value}
-            bind:isValid={item.isValid}
-            bind:startValue={item.valueString}
-            bind:currentValue={item.newValueString}
-            possibleValues={item.possibleValues}
-          />
-        {/if}
-        <!-- {#if item.type == "CUSTOM"}
-            <p>{JSON.stringify(item.customValues)}</p>
-          {/if}
-          {#if item.type == "ARRAY"}
-            <p>{JSON.stringify(item.arrayValues)}</p>
-          {/if} -->
-        <!-- <p>
-            {item.valueString}
-          </p>
-          {JSON.stringify(item)} -->
+        <Field {item} />
       {/each}
     </div>
   {/each}
@@ -192,10 +77,4 @@
 </Section>
 
 <style>
-  .item {
-    border-bottom: 1px solid black;
-  }
-  .item.disabled {
-    opacity: 0.5;
-  }
 </style>
